+++ Backend/src/main/java/com/example/project/dto/response/UserInfoDTO.java
@@ -18,7 +18,7 @@ public class UserInfoDTO {
     private String card_number;
     private Integer CVC;
     private LocalDate validity;
-    private Integer user_discount;
+    private Double user_discount;
     private Integer amount_spent;
     private Integer kol_order;
     private Double percent_order;
